update and move everything to one file
