+++ src/store.js
@@ -1,6 +1,5 @@
 import { legacy_createStore as createStore } from "redux";
-import reducer from "./reducer";
-
+import reducer from "./store/configureStore";
 const store = createStore(reducer);
 
 export default store;
